Add files via upload
couldn't test cause the lab computer is being used for testing on another project.  I did however get the Accel,gyro and mag data to output to the serial port this program should then write that data to a CSV file and also plot this data on a graph not sure if the second step will be helpful or even correct until i test it tomorrow.  '
Michael